requirements.txt
python-telegram-bot==20.7
notion-client==2.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.1.0
openai==1.10.0 
SQLAlchemy==2.0.27 